refactor: application 예외 클래스 이동
