+++ application/src/main/java/com/estime/exception/InternalLogOnlyException.java
@@ -1,6 +1,6 @@
-package com.estime.common.exception.application;
+package com.estime.exception;
 
-import com.estime.common.exception.util.ExceptionMessageFormatter;
+import com.estime.common.util.ExceptionMessageFormatter;
 
 public class InternalLogOnlyException extends ApplicationException {
