+++ core/src/test/java/com/estime/room/domain/participant/ParticipantTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.estime.common.DomainTerm;
-import com.estime.common.exception.domain.InvalidLengthException;
+import com.estime.common.exception.InvalidLengthException;
 import com.estime.room.participant.Participant;
 import com.estime.room.participant.ParticipantName;
 import org.junit.jupiter.api.DisplayName;
