+++ core/src/test/java/com/estime/room/domain/vo/DateSlotTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import com.estime.common.exception.domain.NullNotAllowedException;
+import com.estime.common.exception.NullNotAllowedException;
 import com.estime.room.timeslot.DateSlot;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
