+++ core/src/test/java/com/estime/room/domain/vo/DateTimeSlotTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import com.estime.common.exception.domain.NullNotAllowedException;
-import com.estime.common.exception.domain.SlotNotDivideException;
+import com.estime.common.exception.NullNotAllowedException;
+import com.estime.common.exception.SlotNotDivideException;
 import com.estime.room.timeslot.DateTimeSlot;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
