+++ core/src/test/java/com/estime/room/domain/vo/TimeSlotTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import com.estime.common.exception.domain.InvalidTimeDetailException;
-import com.estime.common.exception.domain.NullNotAllowedException;
-import com.estime.common.exception.domain.SlotNotDivideException;
+import com.estime.common.exception.InvalidTimeDetailException;
+import com.estime.common.exception.NullNotAllowedException;
+import com.estime.common.exception.SlotNotDivideException;
 import com.estime.room.timeslot.TimeSlot;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
